requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pyserial
pyyaml
pyvisa
pyvisa-py